update pooled testing

+++ seirsplus/pooled_test.py
@@ -71,7 +71,7 @@ class OneStageGroupTesting:
         if num_positives:
             sensitivity = num_identified / num_positives
         else:
-            sensitivity = float('nan')
+            sensitivity = float("nan")
         return res, {
             "sensitivity": sensitivity,
             "num_tests": num_tests,
@@ -105,16 +105,12 @@ def run_one_PCR_test(
     LoD = params["LoD"]
 
     if individual:
-        N_templates = np.random.binomial(
-            V_sample * np.array(mu).astype(int), c_1 * xi * c_2
-        )
+        N_templates = np.random.binomial(V_sample * np.array(mu), c_1 * xi * c_2)
         return list(N_templates >= LoD)
 
     pool_size = len(mu)
     # copies of RNA in a subsample that is c_1 of the original volume
-    N_subsamples = np.random.binomial(
-        V_sample * np.array(mu).astype(int), c_1 / pool_size
-    )
+    N_subsamples = np.random.binomial(V_sample * np.array(mu), c_1 / pool_size)
     N_pre_extraction = np.sum(N_subsamples)
     # copies of extracted RNA in the PCR template
     N_templates = np.random.binomial(N_pre_extraction, xi * c_2)
